Added init submodules script
